refactor: split technology slot card locked chrome part (#3878)
Extract LockedResearchSlotCard and shared _SlotCardChrome into a
dedicated library part; document paths in repo-and-packages.

Refs #3878
